+++ src/ir_app/services/__init__.py
@@ -8,6 +8,7 @@ from .evaluation_service import EvaluationService
 from .feedback_analytics_service import FeedbackAnalyticsService
 from .feedback_service import FeedbackService
 from .learning_to_rank_feature_service import LearningToRankFeatureService
+from .learning_to_rank_training_service import LearningToRankTrainingService
 from .ranking_diagnostics_service import RankingDiagnosticsService
 from .retrieval_orchestrator import RetrievalOrchestrator
 from .search_log_service import SearchLogService
@@ -23,6 +24,7 @@ __all__ = [
     "FeedbackService",
     "FeatureUnavailableError",
     "LearningToRankFeatureService",
+    "LearningToRankTrainingService",
     "RankingDiagnosticsService",
     "RetrievalOrchestrator",
     "SearchLogService",